--- src/main/java/ladder/view/output/LadderGameResultOutputView.java
@@ -1,17 +1,0 @@
-package ladder.view.output;
-
-public class LadderGameResultOutputView {
-
-    public static void inputResultPersonName() {
-        System.out.println("결과를 보고 싶은 사람은?");
-    }
-
-    public static void resultGuide() {
-        System.out.println("실행 결과");
-    }
-
-    public static void result(String result) {
-        resultGuide();
-        System.out.println(result);
-    }
-}
